[FIX] summary images may cause shape mismatch [FIX] rename Rsr to Crdn
